fix(membrane): further k-theory runtime math refinements and beads sync
- ktheory.rs: Additional stability improvements for runtime math
  k-theory calculations from concurrent agent iteration
- Beads issue tracker sync with latest session progress

+++ crates/frankenlibc-membrane/src/runtime_math/ktheory.rs
@@ -389,7 +389,7 @@ mod tests {
         let mut ctrl = KTheoryController::new();
         // Need 64 observations per family for baseline freeze.
         // 9 families × 64 = 576 minimum.
-        let total = (CALIBRATION_THRESHOLD as u64) * (NUM_FAMILIES as u64);
+        let total = CALIBRATION_THRESHOLD * NUM_FAMILIES as u64;
         for i in 0..total {
             let family = (i % NUM_FAMILIES as u64) as usize;
             ctrl.observe(family, [0.1, 0.0, 0.0, 0.1]);
